Application Facade and login fuctionality

+++ MultiArmBandit/src/test/TestCumulativeReward.java
@@ -4,8 +4,8 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
-import com.intecs.tddmab.CumulativeReward;
-import com.intecs.tddmab.Reward;
+import com.intecs.mab.CumulativeReward;
+import com.intecs.mab.Reward;
 
 class TestCumulativeReward {
 	private CumulativeReward cumulativeReward() {
@@ -14,7 +14,7 @@ class TestCumulativeReward {
 	}
 
 	private void assertRewardValue(int i, CumulativeReward cumulativeReward) {
-		assertEquals(Integer.valueOf(i), cumulativeReward.getValue());
+		assertEquals(Double.valueOf(i), cumulativeReward.getValue());
 	}
 
 	private Reward reward(int i) {
@@ -36,43 +36,43 @@ class TestCumulativeReward {
 	void testAddUnitReward1() {
 		Reward reward = new Reward(1);
 		CumulativeReward cumulativeReward = new CumulativeReward();
-		assertEquals(Integer.valueOf(0), cumulativeReward.getValue());
+		assertEquals(Double.valueOf(0), cumulativeReward.getValue());
 		cumulativeReward.addReward(reward);
-		assertEquals(Integer.valueOf(1), cumulativeReward.getValue());
+		assertEquals(Double.valueOf(1), cumulativeReward.getValue());
 	}
 	
 	@Test
 	void testAddUnitReward2() {
 		
 		CumulativeReward cumulativeReward = new CumulativeReward(10d);
-		assertEquals(Integer.valueOf(10), cumulativeReward.getValue());
+		assertEquals(Double.valueOf(10), cumulativeReward.getValue());
 		
 		cumulativeReward.addReward(new Reward(1));
-		assertEquals(Integer.valueOf(11), cumulativeReward.getValue());
+		assertEquals(Double.valueOf(11), cumulativeReward.getValue());
 	}
 
 	@Test
 	void testAddZeroReward1() {
 		Reward reward = new Reward(0);
 		CumulativeReward cumulativeReward = new CumulativeReward(0d);
-		assertEquals(Integer.valueOf(0), cumulativeReward.getValue());
+		assertEquals(Double.valueOf(0), cumulativeReward.getValue());
 		cumulativeReward.addReward(reward);
-		assertEquals(Integer.valueOf(0), cumulativeReward.getValue());
+		assertEquals(Double.valueOf(0), cumulativeReward.getValue());
 	}
 
 	@Test
 	void testAddZeroReward2() {
 		Reward reward = new Reward(0);
 		CumulativeReward cumulativeReward = new CumulativeReward(10d);
-		assertEquals(Integer.valueOf(10), cumulativeReward.getValue());
+		assertEquals(Double.valueOf(10), cumulativeReward.getValue());
 		cumulativeReward.addReward(reward);
-		assertEquals(Integer.valueOf(10), cumulativeReward.getValue());
+		assertEquals(Double.valueOf(10), cumulativeReward.getValue());
 	}
 
 	@Test
 	void testResetReward() {
 		CumulativeReward cumulativeReward = new CumulativeReward(10d);
-		assertEquals(Integer.valueOf(10), cumulativeReward.getValue());
+		assertEquals(Double.valueOf(10d), cumulativeReward.getValue());
 		
 		cumulativeReward.reset();
 		assertTrue(cumulativeReward.isZero());
